pyproject.toml
[project]
name = "quiz-prep2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Abdullah-1121", email = "muhmmad11abdullah21@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "langsmith>=0.4.1",
    "openai-agents>=0.0.18",
]

[project.scripts]
quiz-prep2 = "quiz_prep2:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
